+++ libs/api/user/src/lib/domain/services/user.service.ts
@@ -1,40 +1,57 @@
-import { Inject, Injectable } from '@nestjs/common';
-
-import { DeleteResult, UpdateResult } from 'typeorm';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 
 import { CreateUserDTO, UpdateUserDTO, User } from '@angular-auth/libs/common';
 
-import { USER_DRIVEN_ADAPTER_TOKEN } from '../../adapters';
+import { UserDrivenAdapter } from '../../adapters';
 import { UserDrivenPort, UserDriverPort } from '../../ports';
 
 @Injectable()
 export class UserService implements UserDriverPort {
   constructor(
-    @Inject(USER_DRIVEN_ADAPTER_TOKEN)
+    @Inject(UserDrivenAdapter)
     private readonly userDrivenPort: UserDrivenPort
   ) {}
 
   async createUser(user: CreateUserDTO): Promise<User> {
-    return this.userDrivenPort.createUser(user);
+    const { id } = await this.userDrivenPort.createUser(user);
+    return this.userDrivenPort.getUser(id) as Promise<User>;
   }
 
-  async updateUser(id: number, user: UpdateUserDTO): Promise<UpdateResult> {
-    return this.userDrivenPort.updateUser(id, user);
+  async updateUser(id: number, user: UpdateUserDTO): Promise<User> {
+    const { affected } = await this.userDrivenPort.updateUser(id, user);
+
+    if (affected === 0) {
+      throw new NotFoundException(`User with ID ${id} doesn't exist.`);
+    }
+
+    return this.userDrivenPort.getUser(id) as Promise<User>;
   }
 
-  async deleteUser(id: number): Promise<DeleteResult> {
-    return this.userDrivenPort.deleteUser(id);
+  async deleteUser(id: number): Promise<void> {
+    const existsUser = await this.userDrivenPort.getUser(id);
+
+    if (!existsUser) {
+      throw new NotFoundException(`User with ID ${id} doesn't exist.`);
+    }
+
+    this.userDrivenPort.deleteUser(id);
   }
 
-  async getUser(id: number): Promise<User | null> {
-    return this.userDrivenPort.getUser(id);
+  async getUser(id: number): Promise<User> {
+    const user = await this.userDrivenPort.getUser(id);
+
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} doesn't exist.`);
+    }
+
+    return user;
   }
 
   async getUsers(): Promise<User[]> {
     return this.userDrivenPort.getUsers();
   }
 
-  async getUserByEmail(email: string): Promise<User | null> {
+  getUserByEmail(email: string): Promise<User | null> {
     return this.userDrivenPort.getUserByEmail(email);
   }
 }
